+++ Tests/TuistCoreTests/Graph/Nodes/PrecompiledNodeTests.swift
@@ -23,4 +23,37 @@ final class PrecompiledNodeTests: XCTestCase {
         // Then
         XCTAssertEqual(got, "Alamofire")
     }
+
+    func test_is_dynamic_and_linkable_when_xcframework() {
+        // Given
+        let subject = XCFrameworkNode.test()
+
+        // When
+        let got = subject.isDynamicAndLinkable()
+
+        // Then
+        XCTAssertTrue(got)
+    }
+
+    func test_is_dynamic_and_linkable_when_dynamic_framework() {
+        // Given
+        let subject = FrameworkNode.test(linking: .dynamic)
+
+        // When
+        let got = subject.isDynamicAndLinkable()
+
+        // Then
+        XCTAssertTrue(got)
+    }
+
+    func test_is_dynamic_and_linkable_when_static_framework() {
+        // Given
+        let subject = FrameworkNode.test(linking: .static)
+
+        // When
+        let got = subject.isDynamicAndLinkable()
+
+        // Then
+        XCTAssertFalse(got)
+    }
 }